Revert "Kickstart widget on device unlock"
