+++ app/build.gradle.kts
@@ -10,8 +10,8 @@ android {
         applicationId = "net.diffengine.romandigitalclock"
         minSdk = 21
         targetSdk = 34
-        versionCode = 13
-        versionName = "2.2.1"
+        versionCode = 12
+        versionName = "2.2.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
