--- app/src/main/java/net/diffengine/romandigitalclock/receivers/UserPresenceBroadcastReceiver.java
@@ -1,31 +1,0 @@
-/*
- * UserPresenceBroadcastReceiver.java
- * - This file is part of the Android app RomanDigital
- *
- * Copyright 2025 David Yockey
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package net.diffengine.romandigitalclock.receivers;
-
-import static android.content.Intent.ACTION_USER_PRESENT;
-
-// public access specifiers needed to enable access from TimeDisplayWidget
-// (i.e. a non-subclass outside of the package)
-public class UserPresenceBroadcastReceiver extends DroidBroadcastReceiver {
-    public UserPresenceBroadcastReceiver() {
-        super(ACTION_USER_PRESENT);
-    }
-}
